flash: nor: scqspi: Add scqspi_erase flash erase support
This commit implements the `scqspi_erase()` function in the SCQSPI flash
driver, providing sector erase capability.

The function erases a region of flash memory based on the given address
and size, relative to the flash bank base.

This is required for enabling `flash erase_address_range` and supporting
OpenOCD `flash write_image` workflows that depend on pre-erasing flash
sectors.

Signed-off-by: Gaetan Perrot <gaetan.perrot@spacecubics.com>